--- src/main/java/bossbabies/com/a/dao/BookApiDao.java
@@ -1,10 +1,0 @@
-package bossbabies.com.a.dao;
-
-import bossbabies.com.a.dto.BookDto;
-
-import java.util.List;
-
-public interface BookApiDao {
-    boolean insertData(List<BookDto> dtos);
-    List<BookDto> findAll();
-}
