--- src/main/java/bossbabies/com/a/dao/RegisteredBookDao.java
@@ -1,9 +1,0 @@
-package bossbabies.com.a.dao;
-
-import bossbabies.com.a.dto.RegisteredBookDto;
-
-public interface RegisteredBookDao {
-
-    RegisteredBookDto getRegisteredBookById(String registered_book_id);
-
-}
