+++ src/main/java/bossbabies/com/a/dao/book/BookApiDao.java
@@ -1,0 +1,10 @@
+package bossbabies.com.a.dao;
+
+import bossbabies.com.a.dto.BookDto;
+
+import java.util.List;
+
+public interface BookApiDao {
+    boolean insertData(List<BookDto> dtos);
+    List<BookDto> findAll();
+}
